Initial commit for libetnaviv
This breaks out libetnaviv from the etna_viv source code into a
standalone shared library.  This library can be used by itself
to run graphics commands on systems with the galcore kernel
module enabled.  To compile it you only need the public GPL kernel
headers provided by Vivante and also a gc_abi.h header.  There is
a collection of these headers in a repo here.

https://github.com/etnaviv/galcore_headers

Alternatively you can link this library to libvivhook which allows
userspace logging using the binary libGAL.so library.  This is not
enabled by default and only required for advanced command stream
debugging and development.